Extract join button from gateway card, so that it works on any variant of DAO using QV or GW
